# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/instructions/balance": {
            "get": {
                "description": "查询余额指令",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "identity",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/instructions/bind": {
            "post": {
                "description": "绑定钱包指令",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "identity",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/instructions/transfer": {
            "post": {
                "description": "转账指令",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "identity",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request",
                        "name": "ContentBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Transfer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        },
        "/api/instructions/transfer/check": {
            "get": {
                "description": "转账结果查询指令",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "transfer op",
                        "name": "op",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "request.Transfer": {
            "type": "object",
            "properties": {
                "receiver": {
                    "description": "接收人ID",
                    "type": "string"
                },
                "value": {
                    "description": "转账金额",
                    "type": "string"
                }
            }
        }
    }
}